tiff-to-pdf/requirements.txt
requests==2.25.1
img2pdf==0.4.4
flask==2.0.3
Pillow~=9.1.0
pip~=21.1.2
wheel~=0.36.2
lxml~=4.8.0
Jinja2~=3.1.2
Werkzeug~=2.1.2
itsdangerous~=2.1.2
click~=8.1.3
MarkupSafe~=2.1.1
certifi~=2021.10.8
chardet~=4.0.0
pikepdf~=5.1.2
packaging~=21.3
idna~=2.10
urllib3~=1.26.9
pyparsing~=3.0.9
setuptools~=57.0.0
zipp~=3.8.0
